NEW: Add scanning patterns for 1024 grid size
In order to compare the affect on memory access of the different
scanning trajectories we need to use a larger grid size. These
newly added files have a trajectory for each of the 4 types on a
grid 1024x1024 pixels.

The files are located in `data/scanning_patterns`. They were
generated with `6_high_res_scanning.ipynb`. The fly2D file
was too big so it is split into parts aa and ab. Recombine with
`cat`.